feature 10: subscription check
