+++ tgbot/bot/utils.py
@@ -15,7 +15,11 @@ async def check_subscription(
     update: Update,
     context: ContextTypes.DEFAULT_TYPE
 ):
-    """Функция проверяет наличие подписки у пользователя на выбранную группу"""
+    """
+    Функция проверяет наличие подписки у пользователя на выбранную группу.
+    Для того, чтобы проверка сработала нужно добавить бота в администраторы
+    выбранной группы.
+    """
     member = await context.bot.get_chat_member(
         chat_id=CHANNEL_USERNAME,
         user_id=update.callback_query.from_user.id)
